style: RacingManager (클래스명 Racing, 메서드명, 변수명  moving -> advance)
